Consolidate files.
